requirement.txt
numpy
pandas
streamlit
joblib
scikit-learn
streamlit
pillow
pytesseract
scikit-learn